+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Support lacales with date-fns
 - Month name and year format
+- Month missing `styles.month` prop
 - Reset to current date if not selected on close
 
 ### Changed